CHROMIUM: driver: media: virtio: Add V4L2_CID_MPEG_VIDEO_BITRATE_PEAK.
This CL adds support for the V4L2_CID_MPEG_VIDEO_BITRATE_PEAK control
to the virtio encoder. This control allows configuring the peak bitrate
when encoding a video. The peak bitrate is only used when the bitrate
mode is set to VBR (variable bitrate), and is ignored for CBR (constant
bitrate).

Signed-off-by: David Staessens <dstaessens@google.com>

BUG=b:190336806,b:181514834
TEST=tast run DUT arc.VideoEncodeAccel.h264_192p_i420_vm

Cq-Depend: chromium:2944321
Change-Id: Ifbe6708fb5dc296cebfde6e494ebb54cac33f361
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2944506
Tested-by: David Staessens <dstaessens@chromium.org>
Commit-Queue: David Staessens <dstaessens@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>