CHROMIUM: Bluetooth: hci_h5: Set device_wake GPIO on suspend resume
This is a hack which always assert device_wake GPIO when host is
awake, and de-assert when host goes to sleep. Therefore, the
controller's sleep state will reflect the host.

This is not a long-term solution and even if it is merged, a
corrective patch needs to be submitted later.

This is the 4th solution from b/184716907#comment81.

BUG=b:184716907
TEST=Run AVL test with the following FW files provided in the bug.
     *rtl8822cs_fw_0x05A8_C4C1.bin
     *rtl8822cs_config_HWFC_HostWakeBT_activeH_F.bin

Cq-Depend: chromium:2966829
Signed-off-by: Archie Pusaka <apusaka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2944272
Reviewed-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Change-Id: Ib80d44ede7355acb19f7a2ba6f027e5010a5b554
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2966627