FROMGIT: drm/panel: panel-simple: Stash DP AUX bus; allow using it for DDC
If panel-simple is instantiated as a DP AUX bus endpoint then we have
access to the DP AUX bus. Let's stash it in the panel-simple
structure, leaving it NULL for the cases where the panel is
instantiated in other ways.

If we happen to have access to the DP AUX bus and we weren't provided
the ddc-i2c-bus in some other manner, let's use the DP AUX bus for it.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Lyude Paul <lyude@redhat.com>
Reviewed-by: Linus Walleij <linus.walleij@linaro.org>
Link: https://patchwork.freedesktop.org/patch/msgid/20210611101711.v10.6.I18e60221f6d048d14d6c50a770b15f356fa75092@changeid
(cherry picked from commit cc5a3fc041f0104d97ac61b0423ec533d7a93214
  https://github.com/freedesktop/drm-misc.git drm-misc-next)
Signed-off-by: Sean Paul <seanpaul@chromium.org>

BUG=b:191366271
TEST=Tested with single/multiple displays, suspend/resume, hotplugs on volteer/zork/trogdor. Build tested on others

Change-Id: I35571784d7734ac3464a5bf1c3da2e3bf2b5f8ef