CHROMIUM: iwl7000: mac80211: introduce channel switch disconnect function
Introduce a disconnect function that can be used when a
channel switch error occurs. The channel switch can request to
block the tx, and so, we need to make sure we do not send a deauth
frame in this case.

Change-Id: I9edb14785586344af17644b610ab5be109dcef00
Signed-off-by: Nathan Errera <nathan.errera@intel.com>
iwl7000-tree: 116624b0c9f6716cf759cceb034ef387f43bbe6b