CHROMIUM: PCI: mediatek: PCIe host driver delay refclk
This is a workaround patch, in order to force the clkreq#
signal to remain low so that refclk continues to output
for a period of time. It should be reverted when the RTL8822
official fix (https://crrev.com/c/3026232) is ready.

BUG=b:191177754
TEST=emerge-asurada chromeos-kernel-5_4

Signed-off-by: cheng qizhong <qizhong.cheng@mediatek.com>
Change-Id: Ia5a2344da72630a6ef1ec57554065cc5a2f1dd04
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3004458
Tested-by: Qizhong Cheng <qizhong.cheng@mediatek.corp-partner.google.com>
Reviewed-by: Tzung-Bi Shih <tzungbi@chromium.org>
Commit-Queue: Tzung-Bi Shih <tzungbi@chromium.org>
(cherry picked from commit aed0e03af15412611e209bf52ef0b2189a9ebd78)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3033649
Tested-by: Tzung-Bi Shih <tzungbi@chromium.org>