CHROMIUM: arm64: dts: qcom: Add homestar-rev19 as an alias for -rev1
I believe that the board that I have in front of me is supposed to
report as -rev1, but when I updated the BIOS it reports itself as
-rev19. This is because the newer versions of the BIOS support
tri-stated ID pins and the -rev1 boards were misstuffed.

Add -rev19 as an alias for -rev1 in the device tree.

NOTE: -rev2 is apparently showing up as -rev23 due to a similar
problem, but we don't need to handle that yet since that just gets
handled by the default (newest) homestar device tree.

BUG=b:190027674
TEST=I get the right device tree again

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: I96041cce9d6b5df1ed9f01ba2f58ed5be9ddd0bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3032221
Reviewed-by: Julius Werner <jwerner@chromium.org>
(cherry picked from commit 528a4392f340354d4af9bf4ca44373ae5bcfaa16)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3065204
Reviewed-by: Bob Moragues <moragues@chromium.org>
Commit-Queue: Bob Moragues <moragues@chromium.org>
Tested-by: Bob Moragues <moragues@chromium.org>