CHROMIUM: drm/dp: Don't implement "get" for downstream AUX DP backlight
The upstream implementation of this driver doesn't implement the
get(). It's not needed. Not only is it not needed, but it actually
casues confusion too. Specifically on Samsung ATNA33XC20 the "get" was
returning a value that was consistently the value set divided by
32. It appears that when you write to the panel and then get back the
brightness it returns things in a different scale (!) on this panel.

Without being a total expert on DP AUX backlights, this seems like a
bug/quirk in the panel, but since we don't need get() I guess we can
just ignore it.

BUG=b:194652869, b:193287273
TEST="actual_brightness" in sysfs isn't confusing anymore

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: I06ae17e13e74745d177d33c2266fd37665d3442f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3057673
Reviewed-by: Sean Paul <seanpaul@chromium.org>
(cherry picked from commit 0dd27032ae89f04c85a72db4e8b128adb519bddb)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3061360
Commit-Queue: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Bot-Commit: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>