CHROMIUM: arm64: dts: rockchip: add RK3399 gpu-power-model
See also https://crrev.com/c/3078997 for the latest parameters.

Note that Rockchip attempted upstreaming similar bindings previously,
but they were rejected:

  [PATCH v2 5/5] arm64: dts: rockchip: update the thermal zones for RK3399 SoCs
  https://lore.kernel.org/linux-devicetree/1500279271-15249-6-git-send-email-wxt@rock-chips.com/

So we carry this as CHROMIUM, for the never-going-upstream ARM driver.

Also, as in https://crrev.com/c/3078997, we use values used from the
initial release of the Gru family of products, rather than the values in
the above mailing list post.

BUG=b:195781159, b:194727207
TEST=dmesg no longer shows "no DT entry" (from
     kbase_ipa_model_add_param_s32()) complaints
TEST=webGL aquarium stress

Change-Id: I289529060c4fccced922c99ea2e9cfd5ad3edce4
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3086921
Reviewed-by: Chen-Yu Tsai <wenst@chromium.org>