CHROMIUM: media: uvcvideo: add a quirk for ROI fixup
One of the UVC modules that we have returns an error when we try
to `v4l2-ctl -d /dev/videoX -c region_of_interest_auto=1`. The problem
boils down to the fact that default (current) ROI configuration is
invalid - rectangle coordinates outside of GET_MIN. So when we set
auto-controls to 1 what UVC passes to the firmware is invalid ROI
rectangle and valid auto-controls.

Add a fixup function that will check if the device has UVC_QUIRK_ROI
quirk bit set and hence requires ROI configuration fixup.

BUG=b:196800810
TEST=loaded uvc on brya and observed no changes (quirk bit is not set yet)

Signed-off-by: Sergey Senozhatsky <senozhatsky@chromium.org>
Change-Id: I167b8fd71b340dd7cdc5fed8ae2ef38e7f13ba0b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3096740
Reviewed-by: Ricardo Ribalda <ribalda@chromium.org>

[rebase515(groeck):
  Squashed:
    FIXUP: CHROMIUM: media: uvcvideo: add a quirk for ROI fixup
]

Signed-off-by: Guenter Roeck <groeck@google.com>