ANDROID: dma-buf: Add support for mapping buffers with DMA attributes
This reverts commit 9240b79354edf05a2ff5750c0be67bbee27cefe1.

While the dma_map_attrs field is unused in the AOSP kernel, vendor
modules can use this field when mapping DMA buffers. For example, a driver
can use this field to pass in the DMA_ATTR_SKIP_CPU_SYNC DMA attribute
and have it propagate to the DMA layer so that it does not perform cache
maintenance when mapping the buffer.

This is particularly useful in the case where a driver may want to use
the dma_buf_[begin/end]_cpu_access_partial functions to apply cache
maintenance to a subset of the buffer, as opposed to the entire buffer.

[CPNOTE: 06/09/21] Lee: Required by vendor - upstream?

Bug: 197668828
Change-Id: I6dad6b1f9090e92b491d2c88e50b638e34821972
Signed-off-by: Isaac J. Manjarres <isaacm@codeaurora.org>