CHROMIUM: Enable CONFIG_WERROR only for important test builds
If CONFIG_WERROR is enabled on test builds for all architectures, 0-day would
flood us with irrelevant errors. Only enable it for configurations which are
important for us, ie ARM64, and X86.

We should also enable CONFIG_WERROR for arm builds, but that is probably
hopeless due to several frame_size errors observed with clang builds.

BUG=b:199583119
TEST=Test builds

Change-Id: I8a8b656b61827e6eb38187a37ca2c34f58ff6577
Disallow-Recycled-Builds: test-failures
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3156339
Reviewed-by: Marty Faltesek <mfaltesek@google.com>
Reviewed-by: William K Lin <wklin@chromium.org>