FROMGIT: drm/panel: boe-tv101wum-nl6: Support enabling a 3.3V rail
The auo,b101uan08.3 panel (already supported by this driver) has
a 3.3V rail that needs to be turned on. For previous users of
this panel this voltage was directly output by pmic. On a new
user (the not-yet-upstream sc7180-trogdor-mrbland board) we need
to turn the 3.3V rail on. Add support in the driver for this.

Signed-off-by: yangcong <yangcong5@huaqin.corp-partner.google.com>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Link: https://patchwork.freedesktop.org/patch/msgid/20210914032252.3770756-2-yangcong5@huaqin.corp-partner.google.com
(cherry picked from commit 18c58153b8c62f3be87f896274a70778dbd4af9e
 git://anongit.freedesktop.org/drm/drm-misc drm-misc-next)

BUG=b:194156690, b:197555902
TEST=emerge-strongbad chromeos-kernel-5_4

Change-Id: Ia0af5fd5149e347ee3aef003e5a523cdbe98370d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3191096
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>