UPSTREAM: Bluetooth: Fix handling of experimental feature for quality reports
The existence of the experimental feature identifiy is the indication
that it is supported or not. No extra flag needed and the initial flag
should define if a feature is enabled or not. This is actually defined
in the management API definition.

Fixes: ae7d925b5c043 ("Bluetooth: Support the quality report events")
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>
Signed-off-by: Luiz Augusto von Dentz <luiz.von.dentz@intel.com>
(cherry picked from commit 823f3bc4e2eca7170c2367c050c5361cb1cd7c25)

BUG=b:203035116
TEST=Bluetooth quality report works.
For machines that do not support the quality report feature, the
messages log should show

Change-Id: I395563d27239fc69cdf236e73c2f8a0494f541c3
Disallow-Recycled-Builds: test-failures
Signed-off-by: Guenter Roeck <groeck@google.com>