CHROMIUM: usb: dwc3: core: Power off phy in suspend for non-wake-enabled
In this CL [1], Qualcomm-specific wakeup requires keeping the phy
powered. But Rockchip devices don't support wakeup, and in fact need the
phy off. To reconcile the two, look at device_may_wakeup() (matching
xhci-plat.c) to see whether we really intend to support wake.

Given this series is still in the process of upstreaming and has already
diverged quite a bit, let's take this as a CHROMIUM patch for now.

[1] FROMLIST: usb: dwc3: core: Host wake up support from system suspend
    https://crrev.com/c/2352633

BUG=b:163818407, b:196500382
TEST=scarlet suspend/resume + USB

Signed-off-by: Brian Norris <briannorris@chromium.org>
Change-Id: Id77ae8d07aa7266d7d92d94fa25fd163a1cb764e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3192644
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>