CHROMIUM: iwl7000: pcie: make sure prph_info is set when treating wakeup IRQ
In some rare cases when the HW is in a bad state, we may get this
interrupt when prph_info is not set yet.  Then we will try to
dereference it to check the sleep_notif element, which will cause an
oops.

Fix that by ignoring the interrupt if prph_info is not set yet.

Change-Id: I183bb336345b9b3da196ba9e596a6f189fbcbd09
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: d75f4ee62888f596d19dfe9a240543c01c265f34