FROMGIT: ath11k: fix workqueue not getting destroyed after rmmod
Currently, ath11k_core_alloc() creates a single thread workqueue.
This workqueue is not detroyed during clean up when ath11k modules
are unloaded from the kernel and is left as it is.
If workqueue is not destroyed, it could lead to kernel
memory scarcity in a longer run. This could affect self and
other drivers workability as well.

Add destroy workqueue in ath11k_core_free().

Change-Id: I0f6cf952f5737016ee2719d8d095b5f6933e8a2d
Signed-off-by: sampnimm <quic_sampnimm@quicinc.com>