INTERNAL: ath11k: Fix failed to parse regulatory event
In the continuos WiFi up/down scenarios, we see a lot of
"failed to parse regulatory event" prints on the console,
this is beacuse we already would have created a regulatory
domain for the wiphy to use and we will discard the
subsequent events to create the same regulatory event.

Fix this error print by freeing up the regulatory resources
in WiFi down and create the database again in WiFi up.

Change-Id: I24013427152efecbfd1630c9b0ff902404929b46
Signed-off-by: sampnimm <quic_sampnimm@quicinc.com>