INTERNAL: soundwire: qcom: adjust autoenumeration timeout
Currently autoenumeration timeout was set to a very high value of 2 secs which
is really a big value. This can have an adverse effect on boot time if the slave
device is not ready/reset. This was the case with wcd938x which was not reset
yet but we spent 2 secs waiting in the soundwire controller probe which is not
good. Reduce this time to 1/10 of Hz which should be good enough time to finish
autoenumeration if any slaves are available on the bus.

Change-Id: I351e7ceadcddecda1797864fdbdea8d372fdb485
Signed-off-by: sampnimm <quic_sampnimm@quicinc.com>