FROMGIT: ath11k: PCI changes to support WCN6750
In order to add the support for WCN6750 in ATH11K , it is
required to move certain PCI definitions to the header file.
As a result, add ATH11K_PCI_* prefix to these definitions.

Also, change the scope of certain PCI APIs that are required
to enable WCN6750 from static to global.

Change-Id: I9223658e8f7adf68e7d58f1a6e8ef1bb28151902
Signed-off-by: sampnimm <quic_sampnimm@quicinc.com>