FROMGIT: ath11k: Update WBM idle ring HP after FW mode on
Currently, WBM idle ring HP is updated much before the shadow
configuration is sent to the FW. Any update to the shadow
registers before FW mode on request would not be reflected
on to the actual HW registers failing to bring up the device.
Send FW mode ON QMI request before WBM idle ring HP update
to fix this problem.

Change-Id: Icdc099952a6d071bcd48c8b5e16c197c2849b131
Signed-off-by: sampnimm <quic_sampnimm@quicinc.com>