INTERNAL: ath11k: Fix race between CE tasklet and device suspend
In the current code, there is a corner case possiblity that the
CE tasklets can race with the device suspend logic. Due to the
race, if the device registers (HP, TP) are accessed after the
device has suspended, it would result in system faults.

Fix this by running the tasklets to completion before suspending
the device.

Change-Id: I7e1269a69c455feab6ac559a29d08fb6d7e281fe
Signed-off-by: sampnimm <quic_sampnimm@quicinc.com>