BACKPORT: FROMGIT: ath11k: Add HW params for WCN6750
WCN6750 is a PCIe based solution that is attached to and enumerated
by the WPSS (Wireless Processor SubSystem) Q6 processor.

Though it is a PCIe device, since it is not attached to APSS processor
(Application Processor SubSystem), APSS will be unaware of such a decice
and hence it is registered to the APSS processor as a platform device(AHB).
Because of this hybrid nature, it is called as a hybrid bus device.

A new variable hybrid_bus_type is defined in hw_params to indicate the
hybrid nature of the device.

Add HW params for WCN6750.

Change-Id: Ie46b5311b609466c3ed8c5bb73cb3ad3ae2aa010
Signed-off-by: sampnimm <quic_sampnimm@quicinc.com>