FROMGIT: ath11k: Datapath changes to support WCN6750
HAL RX descriptor for WCN6750 is same as QCN9074, this
means that the size of the HAL RX decriptor and the DP
APIs that WCN6750 requires to enable datapath should be
initialized with that of QCN9074's RX descriptor size
and the DP APIs respectively. There is one change wrt to
REO configuration though, REO configuration for WCN6750
follows WCN6855, therefore use reo_setup() of WCN6855
for WCN6750.

Change-Id: I5d5ea742c0ebb59bf6dc1b16db4a4c236f9e952f
Signed-off-by: sampnimm <quic_sampnimm@quicinc.com>