FROMLIST: remoteproc: sysmon: Wait for SSCTL service to come up
The SSCTL service comes up after a finite time when the remote Q6 comes
out of reset. Any graceful shutdowns requested during this period will
be a NOP and abrupt tearing down of the glink channel might lead to pending
transactions on the remote Q6 side and will ultimately lead to a fatal
error. Fix this by waiting for the SSCTL service when a graceful shutdown
is requested.

Change-Id: I9b403ce5c43105122369e861346fd99a119e8aa0
Signed-off-by: sampnimm <quic_sampnimm@quicinc.com>