INTERNAL: ath11k: Init hw_params before setting up AHB resources
As part of adding the support of WCN6750 to ath11k, bus_params
were moved to hw_params and this regressed the initialization
of WCN6750. By the time AHB resources are setup for WCN6750,
hw_params will not be intialized and therefore initialization
for WCN6750 will fail. Fix this by moving the initialization
of hw_params before setting up AHB resources.

Change-Id: Ibe3e618d395962d6495e5945b2aa61d6de1bace6
Signed-off-by: sampnimm <quic_sampnimm@quicinc.com>